+++ switcher/gst-video-converter.cpp
@@ -30,30 +30,74 @@ SWITCHER_MAKE_QUIDDITY_DOCUMENTATION(GstVideoConverter,
                                      "LGPL",
                                      "Nicolas Bouillot");
 
-GstVideoConverter::GstVideoConverter(const std::string&) : shmcntr_(static_cast<Quiddity*>(this)) {}
-
-bool GstVideoConverter::init() {
-  converter_ = std::make_unique<GstPixelFormatConverter>(
-      static_cast<Quiddity*>(this), "Pixel format", "Convert to selected pixel format");
+GstVideoConverter::GstVideoConverter(const std::string&)
+    : video_format_(GstPixelFormatConverter::get_formats(), 0),
+      video_format_id_(
+          pmanage<MPtr(&PContainer::make_selection)>("Pixel format",
+                                                     [this](const size_t& val) {
+                                                       video_format_.select(val);
+                                                       return true;
+                                                     },
+                                                     [this]() { return video_format_.get(); },
+                                                     "Convert to selected pixel format",
+                                                     "Pixel format to convert into",
+                                                     video_format_)),
+      shmcntr_(static_cast<Quiddity*>(this)) {
   shmcntr_.install_connect_method(
       [this](const std::string& shmpath) { return this->on_shmdata_connect(shmpath); },
       [this](const std::string&) { return this->on_shmdata_disconnect(); },
       [this]() { return this->on_shmdata_disconnect(); },
       [this](const std::string& caps) { return this->can_sink_caps(caps); },
       1);
+}
+
+bool GstVideoConverter::init() {
+  shmpath_converted_ = make_file_name("video");
   return true;
 }
 
-bool GstVideoConverter::on_shmdata_disconnect() { return converter_->stop(); }
+bool GstVideoConverter::on_shmdata_disconnect() {
+  prune_tree(".shmdata.writer." + shmpath_converted_);
+  prune_tree(".shmdata.reader." + shmpath_to_convert_);
+  shmsink_sub_.reset();
+  shmsrc_sub_.reset();
+  converter_.reset(nullptr);
+  pmanage<MPtr(&PContainer::enable)>(video_format_id_, true);
+  return true;
+}
 
 bool GstVideoConverter::on_shmdata_connect(const std::string& shmpath) {
-  return converter_->start(shmpath, make_file_name("video-converted"));
+  shmpath_to_convert_ = shmpath;
+  converter_.reset(nullptr);
+  converter_ = std::make_unique<GstPixelFormatConverter>(
+      shmpath_to_convert_, shmpath_converted_, video_format_.get_current_nick());
+  if (!static_cast<bool>(*converter_.get())) return false;
+  shmsink_sub_ = std::make_unique<GstShmdataSubscriber>(
+      converter_->get_shmsink(),
+      [this](const std::string& caps) {
+        graft_tree(".shmdata.writer." + shmpath_converted_,
+                   ShmdataUtils::make_tree(caps, ShmdataUtils::get_category(caps), 0));
+      },
+      [this](GstShmdataSubscriber::num_bytes_t byte_rate) {
+        graft_tree(".shmdata.writer." + shmpath_converted_ + ".byte_rate",
+                   InfoTree::make(byte_rate));
+      });
+  shmsrc_sub_ = std::make_unique<GstShmdataSubscriber>(
+      converter_->get_shmsrc(),
+      [this](const std::string& caps) {
+        graft_tree(".shmdata.reader." + shmpath_to_convert_,
+                   ShmdataUtils::make_tree(caps, ShmdataUtils::get_category(caps), 0));
+      },
+      [this](GstShmdataSubscriber::num_bytes_t byte_rate) {
+        graft_tree(".shmdata.reader." + shmpath_to_convert_ + ".byte_rate",
+                   InfoTree::make(byte_rate));
+      });
+  pmanage<MPtr(&PContainer::enable)>(video_format_id_, false);
+  return true;
 }
 
 bool GstVideoConverter::can_sink_caps(const std::string& caps) {
-  // assuming codecs_ is internally using videoconvert as first caps negotiating
-  // gst element:
-  return GstUtils::can_sink_caps("videoconvert", caps);
+  return GstPixelFormatConverter::can_sink_caps(caps);
 }
 
 }  // namespace switcher